+++ src/command_modules/azure-cli-network/HISTORY.rst
@@ -3,7 +3,7 @@
 Release History
 ===============
 
-2.0.13 (2017-08-28)
+unreleased
 +++++++++++++++++++
 * BC `vnet list-private-access-services`: renamed to `vnet list-endpoint-services`
 * BC `vnet subnet create/update`: renamed `--private-access-services` to `--service-endpoints`
@@ -11,6 +11,10 @@ Release History
 * `lb create`: Added support for SKU.
 * `public-ip create`: Added support for SKU.
 
+2.0.13 (2017-08-28)
++++++++++++++++++++
+* `vnet-gateway`: Added commands `list-bgp-peer-status`, `list-learned-routes` and `list-advertised-routes`
+* `vnet-gateway`: Added command `vpn-client generate`.
 
 2.0.12 (2017-08-11)
 +++++++++++++++++++